third_party/blink/web_tests/platform/generic/external/wpt/html/semantics/embedded-content/the-object-element/object-in-object-fallback-2-expected.txt

This is a testharness.js-based test.
FAIL Test <object> nesting inside <object> assert_equals: Should have loaded all should-load elements expected 12 but got 0
Harness: the test ran to completion.
